use R8 main
- indent everything with 2 spaces
- convert convention plugins to regular plugins
