--- build-logic/src/main/kotlin/gr8.build.common.gradle.kts
@@ -1,23 +1,0 @@
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
-plugins {
-    id("org.jetbrains.kotlin.jvm")
-}
-
-repositories {
-    mavenCentral()
-    google()
-}
-
-java {
-    sourceCompatibility = JavaVersion.VERSION_1_8
-    targetCompatibility = JavaVersion.VERSION_1_8
-}
-
-dependencies {
-    compileOnly("org.jetbrains.kotlin:kotlin-stdlib") // let the Gradle plugin decide the version
-}
-
-rootProject.tasks.named("ci") {
-    dependsOn(tasks.named("build"))
-}
