+++ plugin-common/build.gradle.kts
@@ -1,20 +1,20 @@
 plugins {
-    id("gr8.build.common")
-    id("gr8.build.publishing")
+  id("gr8.build.common")
+  id("gr8.build.publishing")
 }
 
 dependencies {
-    api("dev.gradleplugins:gradle-api:6.0")
-    implementation("net.mbonnin.r8:r8:3.2.49")
+  api("dev.gradleplugins:gradle-api:6.0")
+  implementation("net.mbonnin.r8:r8:3.4.0-dev.0")
 }
 
 val name = "Gr8 Plugin Common"
 val description = "The Gr8 Plugin bytecode with no plugin declaration"
 
 gr8Publishing {
-    configurePublications(name, description)
+  configurePublications(name, description)
 }
 
 publishing.publications.create("default", MavenPublication::class.java) {
-    from(components.getByName("java"))
+  from(components.getByName("java"))
 }
