--- plugin-common/src/main/kotlin/com/gradleup/gr8/Patcher.kt
@@ -1,8 +1,0 @@
-package com.gradleup.gr8
-
-import org.objectweb.asm.ClassReader
-import org.objectweb.asm.ClassVisitor
-import org.objectweb.asm.ClassWriter
-import org.objectweb.asm.Opcodes
-
-
