+++ test-plugin/gradle.properties
@@ -1,0 +1,1 @@
+org.gradle.jvmargs=-Xmx8g
