+++ test-plugin/settings.gradle.kts
@@ -1,1 +1,10 @@
-includeBuild("../")
+pluginManagement {
+  includeBuild("../")
+
+  repositories {
+    //mavenLocal()
+    mavenCentral()
+    google()
+    gradlePluginPortal()
+  }
+}
